Break out tests into separate categories
The one test file was getting too long and confusing and there were too
many Stratis and devicemapper dependencies in it.
