+++ tests/util/mod.rs
@@ -2,5 +2,9 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
+
+pub mod blockdev_tests;
+pub mod dm_tests;
+pub mod setup_tests;
 pub mod simple_tests;
 pub mod logger;
